requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
minio
aiofiles
python-multipart
pydantic-settings